+++ src/observer/sql/parser/value.h
@@ -15,6 +15,7 @@ See the Mulan PSL v2 for more details. */
 #pragma once
 
 #include <string>
+#include <cstdint>
 
 /**
  * @brief 属性的类型
@@ -33,8 +34,8 @@ enum AttrType
 const char *attr_type_to_string(AttrType type);
 AttrType attr_type_from_string(const char *s);
 
-bool deserialize_date(char *out, size_t len_out, u_int in);
-bool serialize_date(u_int *out, const char *in);
+bool deserialize_date(char *out, size_t len_out, int in);
+bool serialize_date(int64_t *out, const char *in);
 
 /**
  * @brief 属性的值
@@ -51,7 +52,7 @@ public:
   }
 
   explicit Value(int val);
-  explicit Value(u_int val);
+  explicit Value(int64_t val);
   explicit Value(float val);
   explicit Value(bool val);
   explicit Value(const char *s, int len = 0);
@@ -69,7 +70,7 @@ public:
     this->set_data(const_cast<char *>(data), length);
   }
   void set_int(int val);
-  void set_date(u_int val);
+  void set_date(int64_t val);
   void set_float(float val);
   void set_boolean(bool val);
   void set_string(const char *s, int len = 0);
@@ -96,7 +97,7 @@ public:
    * 如果当前的类型与期望获取的类型不符，就会执行转换操作
    */
   int get_int() const;
-  u_int get_date() const;
+  int64_t get_date() const;
   float get_float() const;
   std::string get_string() const;
   bool get_boolean() const;
@@ -107,7 +108,7 @@ private:
 
   union {
     int int_value_;
-    u_int date_value_;
+    int64_t date_value_;
     float float_value_;
     bool bool_value_;
   } num_value_;
